renaming fixture files
