+++ src/agent0/test_fixtures/__init__.py
@@ -1,5 +1,5 @@
 """Test fixtures for agent0."""
 
-from .chain import chain, clean_chain, init_chain
-from .hyperdrive import clean_hyperdrive, hyperdrive, init_hyperdrive
-from .interface import hyperdrive_read_interface, hyperdrive_read_write_interface
+from .chain_fixture import chain, clean_chain, init_chain
+from .hyperdrive_fixture import clean_hyperdrive, hyperdrive, init_hyperdrive
+from .interface_fixture import hyperdrive_read_interface, hyperdrive_read_write_interface
